+++ cmn/src/main/java/core/aws/task/as/ASGroupHelper.java
@@ -6,12 +6,11 @@ import com.amazonaws.services.autoscaling.model.Ebs;
 import com.amazonaws.services.autoscaling.model.Tag;
 import core.aws.client.AWS;
 import core.aws.env.Environment;
-import core.aws.resource.as.AutoScalingGroup;
+import core.aws.resource.as.ASGroup;
 import core.aws.resource.as.LaunchConfig;
 import core.aws.resource.vpc.SubnetType;
+import core.aws.util.Encodings;
 import core.aws.util.Randoms;
-import core.aws.util.Strings;
-import org.apache.commons.codec.binary.Base64;
 
 /**
  * @author neo
@@ -23,7 +22,7 @@ public class ASGroupHelper {
         this.env = env;
     }
 
-    public void createLaunchConfig(AutoScalingGroup asGroup) {
+    public void createLaunchConfig(ASGroup asGroup) {
         String launchConfigName = env.name + "-" + asGroup.id + "-" + Randoms.alphaNumeric(6);
 
         LaunchConfig launchConfig = asGroup.launchConfig;
@@ -34,7 +33,7 @@ public class ASGroupHelper {
             .withInstanceType(launchConfig.instanceType)
             .withImageId(launchConfig.ami.imageId())
             .withSecurityGroups(launchConfig.securityGroup.remoteSecurityGroup.getGroupId())
-            .withUserData(Base64.encodeBase64String(Strings.bytes(userData(asGroup))));
+            .withUserData(Encodings.base64(userData(asGroup)));
 
         if (asGroup.subnet.type == SubnetType.PUBLIC) {
             request.withAssociatePublicIpAddress(true);
@@ -52,17 +51,17 @@ public class ASGroupHelper {
         launchConfig.remoteLaunchConfig = AWS.as.createLaunchConfig(request);
     }
 
-    private String userData(AutoScalingGroup asGroup) {
+    private String userData(ASGroup asGroup) {
         return "env=" + env.name + '&'
             + "id=" + asGroup.id + '&'
             + "name=" + instanceName(asGroup);
     }
 
-    Tag nameTag(AutoScalingGroup asGroup) {
+    Tag nameTag(ASGroup asGroup) {
         return new Tag().withKey("Name").withValue(instanceName(asGroup)).withPropagateAtLaunch(true);
     }
 
-    String instanceName(AutoScalingGroup asGroup) {
+    String instanceName(ASGroup asGroup) {
         StringBuilder name = new StringBuilder().append(env.name).append(':').append(asGroup.id);
         if (!asGroup.id.equals(asGroup.launchConfig.ami.id())) {
             name.append(':').append(asGroup.launchConfig.ami.id());