+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 7/29/2015 - 7/31/2015 (1.5.6, not release yet)
 * validate ELB name
 * updated AWS sdk to 1.10.8
+* optimized ASG loading, use one request to load all LaunchConfig
 
 ### 7/18/2015 - 7/20/2015 (1.5.5)
 * update elasticsearch role to 1.7